requirements.txt
oracledb
pandas
openpyxl
pyyaml
requests
tqdm